ssl package refactoring
