+++ nexus-security/src/main/java/com/github/nexus/ssl/exception/NexusSecurityException.java
@@ -1,5 +1,5 @@
 
-package com.github.nexus.ssl;
+package com.github.nexus.ssl.exception;
 
 
 public class NexusSecurityException extends RuntimeException {
