--- nexus-security/src/main/resources/META-INF/services/com.github.nexus.ssl.SSLContextFactory
@@ -1,1 +1,0 @@
-com.github.nexus.ssl.DefaultSSLContextFactory
