+++ nexus-security/src/main/resources/META-INF/services/com.github.nexus.ssl.context.SSLContextFactory
@@ -1,0 +1,1 @@
+com.github.nexus.ssl.context.DefaultSSLContextFactory
